tests: rename /internal/services to /internal/generated to clearly mark such code
This will help with more optimized testing and more accurate coverage stats.
